Tests/LibWeb/Text/expected/navigation/intercepted-traverse-updates-history-entry.txt

navigate destination=/r/programming current=/r/programming/comments/post location=/r/programming/comments/post
handler current=/r/programming/comments/post location=/r/programming/comments/post
navigatesuccess current=/r/programming/comments/post location=/r/programming/comments/post
currententrychange current=/r/programming location=/r/programming
